requirements.txt
pymongo==4.11.1
python-dotenv==1.0.1
fastapi==0.115.7
pydantic==2.10.6
uvicorn==0.34.0
jinja2==3.0.2